:art: Upgrade to latest ESLint config
